LayoutTests/inspector/controller/runtime-controller-expected.txt

CONSOLE MESSAGE: %o
CONSOLE MESSAGE: %o
CONSOLE MESSAGE: %o
CONSOLE MESSAGE: Thrown exception
CONSOLE MESSAGE: %o
CONSOLE MESSAGE: Promise.reject
CONSOLE MESSAGE: Rejection
CONSOLE MESSAGE: %o
CONSOLE MESSAGE: %o
Tests for RuntimeManager operations.


== Running test suite: RuntimeManager
-- Running test case: RuntimeManager.prototype.evaluateInInspectedWindow.ObjectLiteralConvenience
Source: {a:1}
PASS: Evaluation should produce an object.
Source: {a:1, b:2}
PASS: Evaluation should produce an object.
Source:    {a:1, b:2}
PASS: Evaluation should produce an object.
Source: { let a = 1; a += 1; a }
PASS: Evaluation should produce an exception.
Source: ;{a:1}
PASS: Evaluation should produce the labeled statement's value.
Source: {a:1};
PASS: Evaluation should produce the labeled statement's value.
Source: ;{a:1, b:2}
PASS: Evaluation should produce an exception.
Source: ;{ let a = 1; a += 1; a }
PASS: Evaluation should produce the labeled statement's value.

-- Running test case: RuntimeManager.prototype.evaluateInInspectedWindow.AwaitConvenience

Source: await 1
PASS: Transformed. Should log the value or an exception.
Source:    await 2
PASS: Transformed. Should log the value or an exception.
Source: var x = await 3
PASS: Transformed. Should log the value or an exception.
Source: await causeExceptionImmediately()
PASS: Transformed. Should log the value or an exception.
Source: await Promise.resolve(4)
PASS: Transformed. Should log the value or an exception.
Source: await Promise.reject('Promise.reject')
PASS: Transformed. Should log the value or an exception.
Source: await rejectedEventually()
PASS: Transformed. Should log the value or an exception.
Source: await asyncOperation()
PASS: Transformed. Should log the value or an exception.
Source: x = await asyncOperation()
PASS: Transformed. Should log the value or an exception.
Source: return 10
PASS: Exception. Should not get transformed and produce a SyntaxError.
Source: await 10; 1
PASS: Exception. Should not get transformed and produce a SyntaxError.
Source: 1; await 10;
PASS: Exception. Should not get transformed and produce a SyntaxError.
Source: x = y = await 10
PASS: Exception. Should not get transformed and produce a SyntaxError.
